Add sketch of specifying option groups
The POSIX Utility argument syntax specifies a single OPTIONS section.
If there are options make sense only with some set of operands
(commands), then it currently seems to be required to list these for
each operand.

This example explores the potential of define multiple option groups,
which would hopefully help both the command line interface designer and
user to specify and select the correct options for each operand.

Signed-off-by: Eero Aaltonen <eero.aaltonen@vaisala.com>